requirements.txt
# データ処理・可視化関連
numpy>=2.2.0
pandas>=2.2.0
matplotlib>=3.10.0
seaborn>=0.13.0
Pillow>=11.0.0
svgwrite>=1.4.3

# 機械学習関連
scikit-learn>=1.6.0
torch>=2.3.0
torchvision>=0.18.0
grad-cam>=1.5.0

# ユーティリティ
tqdm>=4.66.0
PyYAML>=6.0.0
pydantic>=2.10.0

# カスタムパッケージ
confplot>=0.1.0
data_transformer>=0.1.0
